test(postgres): share logical orchestration fixtures
