+++ crates/automata-ci-postgres/tests/store/fixture.rs
@@ -1,0 +1,93 @@
+use automata_ci_core::{JobContentReference, RunId, Sha256Digest, UnixMillis};
+use automata_ci_store::{
+    AdmissionObject, AdmitLogicalWorkflowRun, LogicalActivationObject, LogicalWorkflowInvocationId,
+    LogicalWorkflowJobId, ObjectKey,
+};
+use sha2::{Digest as _, Sha256};
+use uuid::Uuid;
+
+use crate::support::TestResult;
+
+pub(super) fn retime_logical_admission(
+    command: &AdmitLogicalWorkflowRun,
+    admitted_at: UnixMillis,
+) -> TestResult<AdmitLogicalWorkflowRun> {
+    let mut builder = AdmitLogicalWorkflowRun::builder(
+        command.tenant().clone(),
+        command.idempotency().clone(),
+        command.request_digest(),
+        command.repository().clone(),
+        command.workflow_id(),
+        command.workflow_path(),
+        command.workflow_name(),
+        command.git_ref(),
+        command.snapshot_id(),
+        command.source().clone(),
+        command.plan().clone(),
+        command.run_id(),
+        command.run_attempt(),
+        command.root_invocation_id(),
+        command.event_name(),
+        command.event().clone(),
+        command.head_sha().to_vec(),
+        command.jobs().to_vec(),
+        admitted_at,
+    );
+    if let Some(base_context) = command.base_context() {
+        builder = builder.base_context(base_context.clone());
+    }
+    builder = builder.trust_snapshot(command.trust_snapshot().clone());
+    Ok(builder.build()?)
+}
+
+pub(super) fn job_content_reference(object: &AdmissionObject) -> JobContentReference {
+    JobContentReference::new(
+        object.object_key().as_str(),
+        object.digest(),
+        object.encoded_size(),
+        object.media_type(),
+    )
+}
+
+pub(super) fn activation_content_reference(
+    object: &LogicalActivationObject,
+) -> JobContentReference {
+    JobContentReference::new(
+        object.object_key().as_str(),
+        object.digest(),
+        object.encoded_size(),
+        object.media_type(),
+    )
+}
+
+pub(super) fn deterministic_job_id(
+    run_id: RunId,
+    invocation_id: LogicalWorkflowInvocationId,
+    logical_job_id: LogicalWorkflowJobId,
+    matrix_digest: Sha256Digest,
+) -> automata_ci_core::JobId {
+    let mut hasher = Sha256::new();
+    hasher.update(b"automata.workflow-service.logical-job-id.v1\0");
+    hasher.update(run_id.as_uuid().as_bytes());
+    hasher.update(invocation_id.as_uuid().as_bytes());
+    hasher.update(logical_job_id.as_uuid().as_bytes());
+    hasher.update(0_u32.to_be_bytes());
+    hasher.update(1_u32.to_be_bytes());
+    hasher.update(matrix_digest.as_bytes());
+    let digest: [u8; 32] = hasher.finalize().into();
+    let mut bytes = [0_u8; 16];
+    bytes.copy_from_slice(&digest[..16]);
+    bytes[6] = (bytes[6] & 0x0f) | 0x80;
+    bytes[8] = (bytes[8] & 0x3f) | 0x80;
+    automata_ci_core::JobId::from_uuid(Uuid::from_bytes(bytes))
+}
+
+pub(super) fn context_object(key: &str, digest: u8) -> AdmissionObject {
+    AdmissionObject::new(
+        Sha256Digest::from_bytes([digest; 32]),
+        ObjectKey::new(key).expect("context key"),
+        128,
+        "application/vnd.automata.job-runtime-context.protobuf",
+    )
+    .expect("context object")
+}
