+++ crates/automata-ci-postgres/tests/store/mod.rs
@@ -4,6 +4,8 @@ mod contracts;
 #[cfg(feature = "test-support")]
 mod execution;
 #[cfg(feature = "test-support")]
+mod fixture;
+#[cfg(feature = "test-support")]
 mod orchestration;
 #[cfg(feature = "test-support")]
 mod provider;
