+++ crates/automata-ci-postgres/tests/store/orchestration/logical_renewal_ack.rs
@@ -1,13 +1,16 @@
 use crate::github_manifest_fixture;
+use crate::store::fixture::{
+    activation_content_reference, context_object, deterministic_job_id, job_content_reference,
+    retime_logical_admission,
+};
 
 use std::collections::BTreeMap;
 
 use automata_ci_core::{
-    ContextValue, JobAuthorityProfile, JobContentReference, JobExecutionContext,
-    JobInstanceIdentity, JobIr, JobIrEnvelope, JobRuntimeContext, JobSource, RunId,
-    RunValueTemplates, RunnerRequirements, RuntimeBoolean, SemanticStep, Sha256Digest,
-    ShellTemplate, StepId, StepIr, StrategyContext, UnixMillis, ValueTemplate, WorkflowId,
-    WorkflowJobKey,
+    ContextValue, JobAuthorityProfile, JobExecutionContext, JobInstanceIdentity, JobIr,
+    JobIrEnvelope, JobRuntimeContext, JobSource, RunId, RunValueTemplates, RunnerRequirements,
+    RuntimeBoolean, SemanticStep, Sha256Digest, ShellTemplate, StepId, StepIr, StrategyContext,
+    UnixMillis, ValueTemplate, WorkflowId, WorkflowJobKey,
 };
 use automata_ci_store::{
     AcceptManifestPinnedGithubDelivery, AcceptProviderDelivery, ActivatedLogicalInstanceDescriptor,
@@ -2267,7 +2270,7 @@ async fn admit_authenticated_fixture(database: &TestDatabase, fixture: &mut Fixt
         claimed.claimed_at(),
     )
     .await?;
-    fixture.command = logical_command_at(&fixture.command, claimed.claimed_at())?;
+    fixture.command = retime_logical_admission(&fixture.command, claimed.claimed_at())?;
     let authenticated = AuthenticatedGithubDeliveryClaim::new(
         claimed.claim(),
         claimed.attempt(),
@@ -2283,38 +2286,6 @@ async fn admit_authenticated_fixture(database: &TestDatabase, fixture: &mut Fixt
         )
         .await?;
     Ok(())
-}
-
-fn logical_command_at(
-    command: &AdmitLogicalWorkflowRun,
-    admitted_at: UnixMillis,
-) -> TestResult<AdmitLogicalWorkflowRun> {
-    let mut builder = AdmitLogicalWorkflowRun::builder(
-        command.tenant().clone(),
-        command.idempotency().clone(),
-        command.request_digest(),
-        command.repository().clone(),
-        command.workflow_id(),
-        command.workflow_path(),
-        command.workflow_name(),
-        command.git_ref(),
-        command.snapshot_id(),
-        command.source().clone(),
-        command.plan().clone(),
-        command.run_id(),
-        command.run_attempt(),
-        command.root_invocation_id(),
-        command.event_name(),
-        command.event().clone(),
-        command.head_sha().to_vec(),
-        command.jobs().to_vec(),
-        admitted_at,
-    );
-    if let Some(base_context) = command.base_context() {
-        builder = builder.base_context(base_context.clone());
-    }
-    builder = builder.trust_snapshot(command.trust_snapshot().clone());
-    Ok(builder.build()?)
 }
 
 #[allow(clippy::too_many_lines)]
@@ -2373,8 +2344,8 @@ fn prepared_instance(
         execution.workflow_name(),
         execution.git_ref(),
         format!("/runner/work/{namespace}"),
-        content_reference(claimed.event()),
-        activation_reference(&runtime),
+        job_content_reference(claimed.event()),
+        activation_content_reference(&runtime),
     )
     .with_run_id_alias(execution.run_id_alias())
     .with_run_number(execution.run_number())
@@ -2424,46 +2395,6 @@ fn prepared_instance(
     }
 }
 
-fn deterministic_job_id(
-    run_id: RunId,
-    invocation_id: LogicalWorkflowInvocationId,
-    logical_job_id: LogicalWorkflowJobId,
-    matrix_digest: Sha256Digest,
-) -> automata_ci_core::JobId {
-    let mut hasher = Sha256::new();
-    hasher.update(b"automata.workflow-service.logical-job-id.v1\0");
-    hasher.update(run_id.as_uuid().as_bytes());
-    hasher.update(invocation_id.as_uuid().as_bytes());
-    hasher.update(logical_job_id.as_uuid().as_bytes());
-    hasher.update(0_u32.to_be_bytes());
-    hasher.update(1_u32.to_be_bytes());
-    hasher.update(matrix_digest.as_bytes());
-    let digest: [u8; 32] = hasher.finalize().into();
-    let mut bytes = [0_u8; 16];
-    bytes.copy_from_slice(&digest[..16]);
-    bytes[6] = (bytes[6] & 0x0f) | 0x80;
-    bytes[8] = (bytes[8] & 0x3f) | 0x80;
-    automata_ci_core::JobId::from_uuid(Uuid::from_bytes(bytes))
-}
-
-fn content_reference(object: &AdmissionObject) -> JobContentReference {
-    JobContentReference::new(
-        object.object_key().as_str(),
-        object.digest(),
-        object.encoded_size(),
-        object.media_type(),
-    )
-}
-
-fn activation_reference(object: &LogicalActivationObject) -> JobContentReference {
-    JobContentReference::new(
-        object.object_key().as_str(),
-        object.digest(),
-        object.encoded_size(),
-        object.media_type(),
-    )
-}
-
 fn admission_object(key: String, digest: u8, media_type: &str) -> AdmissionObject {
     AdmissionObject::new(
         Sha256Digest::from_bytes([digest; 32]),
@@ -2472,16 +2403,6 @@ fn admission_object(key: String, digest: u8, media_type: &str) -> AdmissionObjec
         media_type,
     )
     .expect("admission object")
-}
-
-fn context_object(key: &str, digest: u8) -> AdmissionObject {
-    AdmissionObject::new(
-        Sha256Digest::from_bytes([digest; 32]),
-        ObjectKey::new(key).expect("context key"),
-        128,
-        "application/vnd.automata.job-runtime-context.protobuf",
-    )
-    .expect("context object")
 }
 
 async fn database_now_ms(database: &TestDatabase) -> TestResult<i64> {
